move timestamps to adaptor base tests
