+++ tests/functional/timestamps/test_timestamps.py
@@ -1,25 +1,5 @@
-import pytest
-from dbt.tests.util import check_relation_has_expected_schema, run_dbt
-
-_MODEL_CURRENT_TIMESTAMP = """
-SELECT {{current_timestamp()}} as current_timestamp,
-       {{current_timestamp_in_utc()}} as current_timestamp_in_utc
-"""
+from tests.adapter.dbt.tests.adapter.basic.test_timestamps import TestCurrentTimestamps
 
 
-class TestCurrentTimestamps:
-    @pytest.fixture(scope="class")
-    def models(self):
-        return {"get_current_timestamp.sql": _MODEL_CURRENT_TIMESTAMP}
-
-    def test_current_timestamps(self, project, models):
-        results = run_dbt(["run"])
-        assert len(results) == 1
-        check_relation_has_expected_schema(
-            project.adapter,
-            relation_name="get_current_timestamp",
-            expected_schema={
-                "current_timestamp": "timestamp with time zone",
-                "current_timestamp_in_utc": "timestamp without time zone",
-            },
-        )
+class TestCurrentTimestampsCore(TestCurrentTimestamps):
+    pass
